Added the functional service for save data from server response

+++ src/main/java/com/sz/tf/stores/services/ITicketRequestService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface ITicketRequestService {
     void addTicketRequest(TicketRequest ticketRequest);
     List<TicketRequest> getTicketRequests();
-    TicketRequest getTicketRequestById(Long id);
+    TicketRequest getTicketRequestById(int id);
 }
